refactor: Board 패키지 이동
